# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "pos-proxy acts as the backend for POS module in the localnetwork\nit handles all the POS operations also handles the offline mode\nscenarios",
    "title": "pos-proxy api",
    "version": "1.0.0"
  },
  "host": "localhost",
  "basePath": "/api/",
  "paths": {
...
    "/api/opera/rooms/": {
      "get": {
        "description": "returns a list of reservations on Opera",
        "tags": [
          "opera"
        ],
        "summary": "List Opera Rooms",
        "operationId": "listRooms",
        "parameters": [
          {
            "name": "inquiry",
            "in": "query",
            "required": true
          },
          {
            "name": "terminal",
            "in": "query",
            "required": true
          },
          {
            "name": "store",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "postListItem",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/postListItem"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "postListItem": {
      "type": "object",
      "properties": {
        "first_name": {
          "type": "string",
          "x-go-name": "FirstName"
        },
        "id": {
          "type": "string",
          "x-go-name": "ReservationId"
        },
        "last_name": {
          "type": "string",
          "x-go-name": "LastName"
        },
        "room_number": {
          "type": "string",
          "x-go-name": "RoomNumber"
        }
      },
      "x-go-name": "PostListItem",
      "x-go-package": "pos-proxy/opera"
    },
...
  }
}